refactor: deduplicate compareVersions, remove loadingMore and dead code
