+++ plugins/backstage-rhaap-common/src/constants.ts
@@ -1,5 +1,19 @@
 export const SCM_INTEGRATION_AUTH_FAILED_CODE = 'INTEGRATION_AUTH_FAILED';
 
+export const compareVersions = (v1: string, v2: string): number => {
+  const parts1 = v1.split('.').map(p => Number.parseInt(p, 10) || 0);
+  const parts2 = v2.split('.').map(p => Number.parseInt(p, 10) || 0);
+  const maxLen = Math.max(parts1.length, parts2.length);
+
+  for (let i = 0; i < maxLen; i++) {
+    const p1 = parts1[i] || 0;
+    const p2 = parts2[i] || 0;
+    if (p1 > p2) return 1;
+    if (p1 < p2) return -1;
+  }
+  return 0;
+};
+
 export const TERMINAL_JOB_STATUSES = new Set([
   'successful',
   'failed',
