+++ plugins/catalog-backend-module-rhaap/src/providers/versionUtils.ts
@@ -1,18 +1,7 @@
 import { Entity } from '@backstage/catalog-model';
+import { compareVersions } from '@ansible/backstage-rhaap-common';
 
-export const compareVersions = (v1: string, v2: string): number => {
-  const parts1 = v1.split('.').map(p => Number.parseInt(p, 10) || 0);
-  const parts2 = v2.split('.').map(p => Number.parseInt(p, 10) || 0);
-  const maxLen = Math.max(parts1.length, parts2.length);
-
-  for (let i = 0; i < maxLen; i++) {
-    const p1 = parts1[i] || 0;
-    const p2 = parts2[i] || 0;
-    if (p1 > p2) return 1;
-    if (p1 < p2) return -1;
-  }
-  return 0;
-};
+export { compareVersions };
 
 export function stampLatestVersionAnnotations(entities: Entity[]): void {
   const groups = new Map<string, Entity>();
