--- plugins/self-service/src/components/CollectionsCatalog/collectionsCache.test.ts
@@ -1,648 +1,0 @@
-import { Entity } from '@backstage/catalog-model';
-import { CatalogApi } from '@backstage/plugin-catalog-react';
-import { collectionsCache } from './collectionsCache';
-
-const mockEntity: Entity = {
-  apiVersion: 'backstage.io/v1alpha1',
-  kind: 'Component',
-  metadata: {
-    name: 'test-collection',
-    uid: 'uid-1',
-    annotations: {
-      'ansible.io/discovery-source-id': 'src-1',
-      'ansible.io/collection-source': 'pah',
-      'ansible.io/collection-source-repository': 'repo1',
-    },
-    tags: ['networking'],
-  },
-  spec: {
-    type: 'ansible-collection',
-    collection_full_name: 'ns.collection',
-    collection_namespace: 'ns',
-    collection_name: 'collection',
-    collection_version: '1.0.0',
-  },
-};
-
-const createMockCatalogApi = (
-  overrides: Partial<CatalogApi> = {},
-): CatalogApi =>
-  ({
-    queryEntities: jest.fn().mockResolvedValue({
-      items: [mockEntity],
-      totalItems: 1,
-    }),
-    ...overrides,
-  }) as unknown as CatalogApi;
-
-describe('collectionsCache', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    collectionsCache.clear();
-  });
-
-  afterEach(() => {
-    collectionsCache.clear();
-  });
-
-  describe('getState', () => {
-    it('returns null when cache is empty', () => {
-      expect(collectionsCache.getState()).toBeNull();
-    });
-
-    it('returns state after updateSyncStatus initializes it', () => {
-      collectionsCache.updateSyncStatus(
-        { 'src-1': { lastSyncTime: '2024-01-01', lastFailedSyncTime: null } },
-        true,
-      );
-      const state = collectionsCache.getState();
-      expect(state).not.toBeNull();
-      expect(state?.hasConfiguredSources).toBe(true);
-    });
-
-    it('returns null when TTL has expired', () => {
-      collectionsCache.updateSyncStatus({}, true);
-
-      // Mock Date.now to simulate TTL expiration
-      const originalDateNow = Date.now;
-      Date.now = jest.fn(() => originalDateNow() + 6 * 60 * 1000); // 6 minutes later
-
-      expect(collectionsCache.getState()).toBeNull();
-
-      Date.now = originalDateNow;
-    });
-  });
-
-  describe('hasData', () => {
-    it('returns false when cache is empty', () => {
-      expect(collectionsCache.hasData()).toBe(false);
-    });
-
-    it('returns false when state exists but has no entities', () => {
-      collectionsCache.updateSyncStatus({}, true);
-      expect(collectionsCache.hasData()).toBe(false);
-    });
-
-    it('returns true after loading entities', async () => {
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-      expect(collectionsCache.hasData()).toBe(true);
-    });
-  });
-
-  describe('isFullyLoaded', () => {
-    it('returns false when cache is empty', () => {
-      expect(collectionsCache.isFullyLoaded()).toBe(false);
-    });
-
-    it('returns true when all entities are loaded', async () => {
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-      expect(collectionsCache.isFullyLoaded()).toBe(true);
-    });
-
-    it('returns false when more entities remain to be loaded', async () => {
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest
-          .fn()
-          .mockResolvedValueOnce({
-            items: [mockEntity],
-            totalItems: 100, // More than initial fetch
-          })
-          .mockResolvedValue({
-            items: [],
-            totalItems: 100,
-          }),
-      });
-
-      await collectionsCache.startLoading(mockCatalogApi);
-      // After initial load with totalItems > loaded, isFullyLoaded should be false
-      // But since background loading continues, we need to check during load
-      const state = collectionsCache.getState();
-      // Since queryEntities returns empty on subsequent calls, it will mark as fully loaded
-      expect(state).not.toBeNull();
-    });
-  });
-
-  describe('isLoading', () => {
-    it('returns false when not loading', () => {
-      expect(collectionsCache.isLoading()).toBe(false);
-    });
-
-    it('returns false after loading completes', async () => {
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-      expect(collectionsCache.isLoading()).toBe(false);
-    });
-  });
-
-  describe('subscribe', () => {
-    it('adds listener and returns unsubscribe function', () => {
-      const listener = jest.fn();
-      const unsubscribe = collectionsCache.subscribe(listener);
-      expect(typeof unsubscribe).toBe('function');
-    });
-
-    it('listener is called when state updates', async () => {
-      const listener = jest.fn();
-      collectionsCache.subscribe(listener);
-
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      expect(listener).toHaveBeenCalled();
-      expect(listener).toHaveBeenCalledWith(
-        expect.objectContaining({
-          entities: expect.arrayContaining([mockEntity]),
-        }),
-      );
-    });
-
-    it('unsubscribe removes the listener', async () => {
-      const listener = jest.fn();
-      const unsubscribe = collectionsCache.subscribe(listener);
-      unsubscribe();
-
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      expect(listener).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('startLoading', () => {
-    it('loads entities from catalogApi', async () => {
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      expect(mockCatalogApi.queryEntities).toHaveBeenCalledWith({
-        filter: { kind: 'Component', 'spec.type': 'ansible-collection' },
-        limit: 50,
-        orderFields: [{ field: 'metadata.uid', order: 'asc' }],
-      });
-
-      const state = collectionsCache.getState();
-      expect(state?.entities).toHaveLength(1);
-      expect(state?.entities[0]).toEqual(mockEntity);
-    });
-
-    it('sets error state when loading fails', async () => {
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest.fn().mockRejectedValue(new Error('API error')),
-      });
-
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      const state = collectionsCache.getState();
-      expect(state?.error).toBe('API error');
-      expect(state?.entities).toHaveLength(0);
-    });
-
-    it('does not restart loading if already loading', async () => {
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest
-          .fn()
-          .mockImplementation(
-            () =>
-              new Promise(resolve =>
-                setTimeout(
-                  () => resolve({ items: [mockEntity], totalItems: 1 }),
-                  100,
-                ),
-              ),
-          ),
-      });
-
-      // Start two loads simultaneously
-      const load1 = collectionsCache.startLoading(mockCatalogApi);
-      const load2 = collectionsCache.startLoading(mockCatalogApi);
-
-      await Promise.all([load1, load2]);
-
-      // Should only be called once for initial load
-      expect(mockCatalogApi.queryEntities).toHaveBeenCalledTimes(1);
-    });
-
-    it('does not reload if already fully loaded', async () => {
-      const mockCatalogApi = createMockCatalogApi();
-
-      await collectionsCache.startLoading(mockCatalogApi);
-      expect(mockCatalogApi.queryEntities).toHaveBeenCalledTimes(1);
-
-      // Try to load again
-      await collectionsCache.startLoading(mockCatalogApi);
-      // Should not call again since already fully loaded
-      expect(mockCatalogApi.queryEntities).toHaveBeenCalledTimes(1);
-    });
-
-    it('continues from partial load when startLoading is called after interrupted background load', async () => {
-      // First load returns partial data (50 of 100 items)
-      const initialEntities = Array.from({ length: 50 }, (_, i) => ({
-        ...mockEntity,
-        metadata: { ...mockEntity.metadata, uid: `uid-${i}` },
-      }));
-
-      const remainingEntities = Array.from({ length: 50 }, (_, i) => ({
-        ...mockEntity,
-        metadata: { ...mockEntity.metadata, uid: `uid-${i + 50}` },
-      }));
-
-      let callCount = 0;
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest.fn().mockImplementation(({ offset }) => {
-          callCount++;
-          if (callCount === 1) {
-            // Initial load - return partial data
-            return Promise.resolve({
-              items: initialEntities,
-              totalItems: 100,
-            });
-          }
-          if (callCount === 2) {
-            // Background load fails - simulates interrupted load
-            return Promise.reject(new Error('Network error'));
-          }
-          // Subsequent calls succeed
-          return Promise.resolve({
-            items: offset === 50 ? remainingEntities : [],
-            totalItems: 100,
-          });
-        }),
-      });
-
-      // First load - gets partial data, background load fails
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      // At this point, cache has partial data (50 items, not fully loaded)
-      // loadingPromise is null because background loading failed and cleared it
-      let state = collectionsCache.getState();
-      expect(state?.loadedOffset).toBe(50);
-      expect(state?.isFullyLoaded).toBe(false);
-      expect(collectionsCache.isLoading()).toBe(false);
-
-      // Call startLoading again - should continue from offset 50
-      // This exercises lines 86-89 (continue from partial data)
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      state = collectionsCache.getState();
-      expect(state?.entities.length).toBe(100);
-      expect(state?.isFullyLoaded).toBe(true);
-    });
-
-    it('waits for existing background load when called again during loading', async () => {
-      // This test covers lines 157-163 where continueBackgroundLoading
-      // waits for an existing loading promise if one is in progress
-      const initialEntities = Array.from({ length: 50 }, (_, i) => ({
-        ...mockEntity,
-        metadata: { ...mockEntity.metadata, uid: `uid-${i}` },
-      }));
-
-      const remainingEntities = Array.from({ length: 150 }, (_, i) => ({
-        ...mockEntity,
-        metadata: { ...mockEntity.metadata, uid: `uid-${i + 50}` },
-      }));
-
-      let backgroundLoadResolve!: (value: unknown) => void;
-      const backgroundLoadPromise = new Promise(resolve => {
-        backgroundLoadResolve = resolve;
-      });
-
-      let callCount = 0;
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest.fn().mockImplementation(({ offset }) => {
-          callCount++;
-          if (callCount === 1) {
-            // Initial load returns partial data
-            return Promise.resolve({
-              items: initialEntities,
-              totalItems: 200,
-            });
-          }
-          // Background load waits for manual resolution
-          return backgroundLoadPromise.then(() => ({
-            items: offset === 50 ? remainingEntities : [],
-            totalItems: 200,
-          }));
-        }),
-      });
-
-      // Start first load - initial load completes, background starts but waits
-      const load1Promise = collectionsCache.startLoading(mockCatalogApi);
-
-      // Give time for initial load to complete
-      await new Promise(resolve => setTimeout(resolve, 10));
-
-      // Verify we have partial data and loading is in progress
-      expect(collectionsCache.getState()?.loadedOffset).toBe(50);
-      expect(collectionsCache.isLoading()).toBe(true);
-
-      // Start second load while first is still loading
-      // This should trigger lines 157-163 (wait for existing load)
-      const load2Promise = collectionsCache.startLoading(mockCatalogApi);
-
-      // Resolve the background load
-      backgroundLoadResolve(undefined);
-
-      // Both promises should resolve
-      await Promise.all([load1Promise, load2Promise]);
-
-      // Should have loaded all data
-      const state = collectionsCache.getState();
-      expect(state?.entities.length).toBe(200);
-      expect(state?.isFullyLoaded).toBe(true);
-    });
-
-    it('handles clear() being called during background loading startup', async () => {
-      // This test covers lines 152-155 where continueBackgroundLoading
-      // returns early if state is cleared before it runs
-      const initialEntities = Array.from({ length: 50 }, (_, i) => ({
-        ...mockEntity,
-        metadata: { ...mockEntity.metadata, uid: `uid-${i}` },
-      }));
-
-      let resolveBackgroundLoad!: () => void;
-      const backgroundLoadPromise = new Promise<void>(resolve => {
-        resolveBackgroundLoad = resolve;
-      });
-
-      let callCount = 0;
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest.fn().mockImplementation(() => {
-          callCount++;
-          if (callCount === 1) {
-            // Initial load returns partial data
-            return Promise.resolve({
-              items: initialEntities,
-              totalItems: 100,
-            });
-          }
-          // Background load waits until we resolve it
-          return backgroundLoadPromise.then(() => ({
-            items: [],
-            totalItems: 100,
-          }));
-        }),
-      });
-
-      // Start loading - initial load completes, background load starts but waits
-      const loadPromise = collectionsCache.startLoading(mockCatalogApi);
-
-      // Give time for initial load to complete and background load to start
-      await new Promise(resolve => setTimeout(resolve, 10));
-
-      // Clear the cache while background loading is pending
-      collectionsCache.clear();
-
-      // Now resolve the background load - it should handle the cleared state gracefully
-      resolveBackgroundLoad();
-
-      await loadPromise;
-
-      // Cache should be cleared
-      expect(collectionsCache.getState()).toBeNull();
-      expect(collectionsCache.isLoading()).toBe(false);
-    });
-
-    it('continues background loading for large datasets', async () => {
-      const entities = Array.from({ length: 60 }, (_, i) => ({
-        ...mockEntity,
-        metadata: { ...mockEntity.metadata, uid: `uid-${i}` },
-      }));
-
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest
-          .fn()
-          .mockResolvedValueOnce({
-            items: entities.slice(0, 50),
-            totalItems: 60,
-          })
-          .mockResolvedValueOnce({
-            items: entities.slice(50),
-            totalItems: 60,
-          }),
-      });
-
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      const state = collectionsCache.getState();
-      expect(state?.entities).toHaveLength(60);
-      expect(state?.isFullyLoaded).toBe(true);
-    });
-  });
-
-  describe('updateSyncStatus', () => {
-    it('creates state if none exists', () => {
-      collectionsCache.updateSyncStatus(
-        { 'src-1': { lastSyncTime: '2024-01-01', lastFailedSyncTime: null } },
-        true,
-      );
-
-      const state = collectionsCache.getState();
-      expect(state).not.toBeNull();
-      expect(state?.syncStatusMap).toEqual({
-        'src-1': { lastSyncTime: '2024-01-01', lastFailedSyncTime: null },
-      });
-      expect(state?.hasConfiguredSources).toBe(true);
-    });
-
-    it('updates existing state', async () => {
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      collectionsCache.updateSyncStatus(
-        { 'src-2': { lastSyncTime: '2024-02-01', lastFailedSyncTime: null } },
-        true,
-      );
-
-      const state = collectionsCache.getState();
-      expect(state?.syncStatusMap).toEqual({
-        'src-2': { lastSyncTime: '2024-02-01', lastFailedSyncTime: null },
-      });
-      expect(state?.entities).toHaveLength(1); // Entities preserved
-    });
-  });
-
-  describe('invalidateFetchedData', () => {
-    it('reloads data without emitting empty state and without dropping listeners', async () => {
-      const listener = jest.fn();
-      const mockCatalogApi = createMockCatalogApi();
-      collectionsCache.subscribe(listener);
-      await collectionsCache.startLoading(mockCatalogApi);
-      listener.mockClear();
-
-      collectionsCache.invalidateFetchedData();
-
-      const emptyCall = listener.mock.calls.find(
-        ([state]: [any]) =>
-          state.entities.length === 0 && state.totalServerItems === 0,
-      );
-      expect(emptyCall).toBeUndefined();
-
-      await new Promise(resolve => setTimeout(resolve, 50));
-
-      expect(collectionsCache.getState()?.entities).toHaveLength(1);
-      expect(listener.mock.calls.length).toBeGreaterThan(0);
-    });
-  });
-
-  describe('clear', () => {
-    it('clears all state', async () => {
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      expect(collectionsCache.hasData()).toBe(true);
-
-      collectionsCache.clear();
-
-      expect(collectionsCache.getState()).toBeNull();
-      expect(collectionsCache.hasData()).toBe(false);
-      expect(collectionsCache.isLoading()).toBe(false);
-    });
-
-    it('clears listeners', async () => {
-      const listener = jest.fn();
-      collectionsCache.subscribe(listener);
-
-      collectionsCache.clear();
-
-      // Update state after clear - listener should not be called
-      collectionsCache.updateSyncStatus({}, true);
-      // updateSyncStatus doesn't notify listeners, so test with startLoading
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      // Listener was cleared, so it shouldn't have been called after clear
-      // Actually the listener is added before clear, so it should be removed
-      // Let's check the state is correct instead
-      expect(collectionsCache.hasData()).toBe(true);
-    });
-  });
-
-  describe('getResumeOffset', () => {
-    it('returns 0 when cache is empty', () => {
-      expect(collectionsCache.getResumeOffset()).toBe(0);
-    });
-
-    it('returns loaded offset after partial load', async () => {
-      const mockCatalogApi = createMockCatalogApi();
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      expect(collectionsCache.getResumeOffset()).toBe(1);
-    });
-  });
-
-  describe('getTotalServerItems', () => {
-    it('returns 0 when cache is empty', () => {
-      expect(collectionsCache.getTotalServerItems()).toBe(0);
-    });
-
-    it('returns total items after load', async () => {
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest.fn().mockResolvedValue({
-          items: [mockEntity],
-          totalItems: 100,
-        }),
-      });
-
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      expect(collectionsCache.getTotalServerItems()).toBe(100);
-    });
-  });
-
-  describe('filters extraction', () => {
-    it('extracts unique sources from PAH entities', async () => {
-      const pahEntity: Entity = {
-        apiVersion: 'backstage.io/v1alpha1',
-        kind: 'Component',
-        metadata: {
-          name: 'collection-1',
-          uid: 'uid-1',
-          tags: ['networking'],
-          annotations: {
-            'ansible.io/collection-source': 'pah',
-            'ansible.io/collection-source-repository': 'my-pah-repo',
-          },
-        },
-        spec: { type: 'ansible-collection' },
-      };
-
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest.fn().mockResolvedValue({
-          items: [pahEntity],
-          totalItems: 1,
-        }),
-      });
-
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      const state = collectionsCache.getState();
-      expect(state?.allSources).toContain('All');
-      expect(state?.allSources).toContain('my-pah-repo');
-      expect(state?.allTags).toContain('All');
-      expect(state?.allTags).toContain('networking');
-    });
-
-    it('extracts unique sources from SCM entities', async () => {
-      const scmEntity: Entity = {
-        apiVersion: 'backstage.io/v1alpha1',
-        kind: 'Component',
-        metadata: {
-          name: 'collection-2',
-          uid: 'uid-2',
-          tags: ['security'],
-          annotations: {
-            'ansible.io/scm-host-name': 'github.com',
-          },
-        },
-        spec: { type: 'ansible-collection' },
-      };
-
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest.fn().mockResolvedValue({
-          items: [scmEntity],
-          totalItems: 1,
-        }),
-      });
-
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      const state = collectionsCache.getState();
-      expect(state?.allSources).toContain('All');
-      expect(state?.allSources).toContain('github.com');
-      expect(state?.allTags).toContain('All');
-      expect(state?.allTags).toContain('security');
-    });
-
-    it('extracts tags and excludes ansible-collection tag', async () => {
-      const entityWithTags: Entity = {
-        apiVersion: 'backstage.io/v1alpha1',
-        kind: 'Component',
-        metadata: {
-          name: 'collection-1',
-          uid: 'uid-1',
-          tags: ['networking', 'ansible-collection', 'cloud'],
-          annotations: {
-            'ansible.io/collection-source': 'pah',
-            'ansible.io/collection-source-repository': 'repo1',
-          },
-        },
-        spec: { type: 'ansible-collection' },
-      };
-
-      const mockCatalogApi = createMockCatalogApi({
-        queryEntities: jest.fn().mockResolvedValue({
-          items: [entityWithTags],
-          totalItems: 1,
-        }),
-      });
-
-      await collectionsCache.startLoading(mockCatalogApi);
-
-      const state = collectionsCache.getState();
-      expect(state?.allTags).toContain('networking');
-      expect(state?.allTags).toContain('cloud');
-      expect(state?.allTags).not.toContain('ansible-collection');
-    });
-  });
-});
